pyproject.toml
[project]
name = "mcp-data-auditor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cosmos>=4.14.6",
    "azure-identity>=1.25.2",
    "faker>=40.4.0",
    "fastmcp>=2.14.5",
    "mcp[cli]>=1.26.0",
    "ruff>=0.15.1",
    "werkzeug>=3.1.5",
]
